chore: map event listener 코드 이동
